Move dashboard to own blueprint containing measure list as well.
Add link to footer.
