+++ application/cms/views.py
@@ -51,83 +51,6 @@ def index():
     return render_template('cms/index.html', pages=pages)
 
 
-@cms_blueprint.route('/measures', methods=['GET'])
-@internal_user_required
-@login_required
-def measures():
-    pages = page_service.get_pages_by_type('topic')
-    return render_template('cms/measures.html', pages=pages)
-
-
-# Temporary measure page fact
-@cms_blueprint.route('/measures/facts-and-figures', methods=['GET'])
-@internal_user_required
-@login_required
-def measure_page_facts_and_figures():
-    from application.cms.models import DbPage
-    from datetime import date, datetime, timedelta
-    import calendar
-
-    measures = DbPage.query.filter(
-        DbPage.publication_date.isnot(None),
-        DbPage.version == '1.0'
-    ).all()
-
-    updates = DbPage.query.filter(
-        DbPage.publication_date.isnot(None),
-        DbPage.version != '1.0'
-    ).all()
-
-    seven_days = timedelta(days=7)
-    seven_days_ago = datetime.today() - seven_days
-    in_last_week = DbPage.query.filter(
-        DbPage.publication_date.isnot(None),
-        DbPage.publication_date >= seven_days_ago
-    ).all()
-
-    first_publication = DbPage.query.filter(
-        DbPage.publication_date.isnot(None)
-    ).order_by(DbPage.publication_date.asc()).first()
-
-    data = {'publications': len(measures),
-            'updates': len(updates),
-            'in_last_week': len(in_last_week),
-            'first_publication': first_publication.publication_date}
-
-    measures_by_week = {}
-
-    def month_iterator(start, end):
-        current = start
-        while current < end:
-            yield current
-            current += timedelta(days=current.max.day)
-
-    def in_range(week, begin, end=date.today()):
-        return any([d for d in week if d >= begin]) and any([d for d in week if d <= end])
-
-    for m in month_iterator(first_publication.publication_date, date.today()):
-        c = calendar.Calendar(calendar.MONDAY).monthdatescalendar(m.year, m.month)
-        for week in c:
-                if in_range(week, first_publication.publication_date):
-                    publications = DbPage.query.filter(
-                        DbPage.publication_date.isnot(None),
-                        DbPage.publication_date >= week[0],
-                        DbPage.publication_date <= week[6],
-                        DbPage.version == '1.0'
-                    ).all()
-                    updates = DbPage.query.filter(
-                        DbPage.publication_date.isnot(None),
-                        DbPage.publication_date >= week[0],
-                        DbPage.publication_date <= week[6],
-                        DbPage.version != '1.0'
-                    ).all()
-                    measures_by_week[week[0]] = {'publications': len(publications), 'updates': len(updates)}
-
-    data['measures_by_week'] = measures_by_week
-
-    return render_template('cms/measure_facts_and_figures.html', data=data)
-
-
 @cms_blueprint.route('/<topic>/<subtopic>/measure/new', methods=['GET', 'POST'])
 @internal_user_required
 @login_required
