+++ application/templates/dashboard/index.html
@@ -6,6 +6,7 @@
         <h1 class='heading-large'>
             Measure facts and figures
         </h1>
+        <p><a href="{{ url_for('dashboard.measures') }}">Find measure pages</a></p>
     </div>
 
     <div class="grid-row">
