+++ application/templates/dashboard/measures.html
@@ -6,9 +6,8 @@
     <div class="grid-row">
         <div class="column-full">
             <h1 class='heading-large no-breadcrumb'>
-                Government data on ethnicity
+                Measure pages
             </h1>
-            <p><a href="{{ url_for('cms.measure_page_facts_and_figures') }}">Measure page facts and figures</a></p>
             <table id="status-table">
 
                 <thead>
